agents/official/beeai-framework/chat/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "AI-powered conversational system with memory and tools."
authors = [{ name = "IBM Corp." }]
requires-python = ">=3.11,<3.12"
dependencies = [
    "acp-sdk>=0.3.3",
    "beeai-framework[duckduckgo,wikipedia]>=0.1.17",
    "pydantic-settings>=2.9.0",
]

[project.scripts]
server = "chat.agent:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chat"]
